+++ packages/dom-snapshot/src/browser/processPage.js
@@ -26,11 +26,12 @@ const makeSessionCache = require('./sessionCache');
 
 function processPage(
   doc = document,
-  {showLogs, useSessionCache, dontFetchResources, fetchTimeout} = {},
+  {showLogs, useSessionCache, dontFetchResources, fetchTimeout, skipResources} = {},
 ) {
   /* MARKER FOR TEST - DO NOT DELETE */
   const log = showLogs ? makeLog(Date.now()) : noop;
   log('processPage start');
+  log(`skipResources length: ${skipResources && skipResources.length}`);
   const sessionCache = useSessionCache && makeSessionCache({log});
   const styleSheetCache = {};
   const extractResourcesFromStyleSheet = makeExtractResourcesFromStyleSheet({styleSheetCache});
@@ -82,7 +83,7 @@ function processPage(
 
     const resourceUrlsAndBlobsPromise = dontFetchResources
       ? Promise.resolve({resourceUrls: urls, blobsObj: {}})
-      : getResourceUrlsAndBlobs({documents: docRoots, urls}).then(result => {
+      : getResourceUrlsAndBlobs({documents: docRoots, urls, skipResources}).then(result => {
           sessionCache && sessionCache.persist();
           return result;
         });